+++ src/utils/getUpdate.js
@@ -69,6 +69,25 @@ function getValue (attr, currentState = {}) {
     return rest ? undefined : state;
 }
 
+function isEmpty (val) {
+    return val === null
+        || val === undefined
+        || `${val}`.trim() === '';
+}
+
+function toArray (previousValue) {
+    if (Array.isArray(previousValue)) {
+        return previousValue.slice();
+    }
+    if (isEmpty(previousValue)) {
+        return [];
+    }
+    return [previousValue];
+}
+
+const ENTITY_HBS_REGEXP = /^\s*\{\{\[?@([^@[\]{}\s]+)(\])?\}\}\s*$/;
+const VARIABLE_HBS_REGEXP = /^\s*\{\{\[?([^@[\]{}\s]+)\]?\}\}\s*$/;
+
 function getSetState (setState = {}, req, res = null, useState = null) {
 
     const keys = Object.keys(setState)
@@ -128,6 +147,50 @@ function getSetState (setState = {}, req, res = null, useState = null) {
                 }
 
                 set = incremet + previousValue;
+            } else if (val._$shift || val._$pop
+                || typeof val._$push !== 'undefined'
+                || typeof val._$add !== 'undefined'
+                || typeof val._$rem !== 'undefined'
+                || typeof val._$set !== 'undefined') {
+
+                let valAsArray = toArray(getValue(k, state));
+
+                if (val._$shift) {
+                    valAsArray.shift();
+                } else if (val._$pop) {
+                    valAsArray.pop();
+                } else {
+                    const value = val._$add || val._$rem || val._$push || val._$set;
+                    const [, entity, rear] = `${value}`.match(ENTITY_HBS_REGEXP) || [];
+                    const [, variable] = `${value}`.match(VARIABLE_HBS_REGEXP) || [];
+
+                    let values;
+                    if (entity && (!rear || req.entities.some((e) => e.entity === entity))) {
+                        values = req.entities.filter((e) => e.entity === entity)
+                            .map((e) => e.value);
+                    } else if (variable) {
+                        values = toArray(getValue(variable, state));
+                    } else {
+                        const useValue = typeof value === 'string'
+                            ? handlebars.compile(value)(stateData(req, res))
+                            : value;
+                        values = toArray(useValue);
+                    }
+
+                    if (typeof val._$push !== 'undefined') {
+                        valAsArray.push(...values);
+                    } else if (typeof val._$add !== 'undefined') {
+                        const unique = Array.from(new Set(values).values());
+                        valAsArray.push(...unique.filter((v) => !valAsArray.includes(v)));
+                    } else if (typeof val._$rem !== 'undefined') {
+                        valAsArray = valAsArray
+                            .filter((v) => !values.includes(v));
+                    } else if (typeof val._$set !== 'undefined') {
+                        valAsArray = values;
+                    }
+                }
+
+                set = valAsArray;
             } else if (k.match(/^(@|_~)/)
                 || !Object.keys(val).some((l) => l.match(/^_\$/))) {
                 set = val;