+++ CHANGELOG.md
@@ -5,11 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.20.0] 2021-09-21
+
+## Added
+
+- support for setState array operators (`_$push`, `_$pop`, `_$shift`, `_$add`, `_$set` `_$rem`)
+
+### Changed
+
+- when using range comparison operators `>`, `<`, `>=`, `<=` against array, it's length is used
+### Fixed
+
+- data param of `res.addQuickReply()` method
 ## [3.19.3] 2021-09-06
 
 ### Fixed
 
 - AI Entity context when using optional entity
+
 ## [3.19.3] 2021-09-06
 
 ### Fixed